backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.9
pydantic-settings>=2.3.4
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.9
email-validator>=2.2.0
pandas>=2.2.2
numpy>=1.26.4
openpyxl>=3.1.5
plotly>=5.22.0
scikit-learn>=1.5.1
reportlab>=4.2.2
